Corrigindo testes

+++ Comex/Comex/Testes/TestaEntradaDeProdutoNoEstoque.cs
@@ -11,11 +11,11 @@ namespace Comex.Testes
         Categoria saude = new Categoria("SAUDE", "ATIVA");
 
         Produto vacina = new ProdutoIsento("Vacina", "Vacinas para todas as doenças",
-            250.00, 500, saude);
+            (decimal)250.00, 500, saude);
         Produto livroJava = new ProdutoIsento("Use a Cabeça: Java", "Descomplicando Java",
-            112.90, 5, livros);
+            (decimal)112.90, 5, livros);
         Produto macbookPro = new ProdutoIsento("Macbook pro", "O melhor da marca!",
-            20000.00, 15, informatica);
+            (decimal)20000.00, 15, informatica);
 
         Estoque estoque = new Estoque();
         estoque.RegistraEntrada(vacina);
@@ -26,7 +26,7 @@ namespace Comex.Testes
 
         Categoria alimentos = new Categoria("ALIMENTOS", "ATIVA");
         Produto bala = new Produto("Bala 7 Belo", "A mais gostosa",
-            0.1, 5000, alimentos);
+            (decimal)0.1, 5000, alimentos);
 
         try
         {
