+++ Comex/Comex/Testes/TestaProduto.cs
@@ -11,11 +11,11 @@ namespace Comex.Testes
         Categoria livros = new Categoria("LIVROS", "ATIVA");
 
         Produto notebook = new Produto("Notebook Samsung", "O melhor da marca!",
-            3523.00, 1, informatica);
+            (decimal)3523.00, 1, informatica);
         Produto cleanA = new Produto("Clean Architecture", "O melhor da marca!",
-            102.90, 2, livros);
+            (decimal)102.90, 2, livros);
         Produto monitor = new Produto("Monitor Dell", "O melhor da marca!",
-            1889.00, 3, informatica);
+            (decimal)1889.00, 3, informatica);
 
         Console.WriteLine($"Produtos: \n\nCategoria: {notebook.Categoria.Nome} \nId: {notebook.Id} - {notebook.Nome}, {notebook.Descricao} " +
             $"R${notebook.Preco} \nQuantidade em estoque: {notebook.Estoque} \nImposto R${notebook.CalculaImposto()}");
