+++ Comex/Comex/Testes/TestaProdutoIsento.cs
@@ -11,11 +11,11 @@ namespace Comex.Testes
         Categoria saude = new Categoria("SAUDE", "ATIVA");
 
         Produto vacina = new ProdutoIsento("Vacina", "Vacinas para todas as doenças",
-            250.00, 500, saude);
+            (decimal)250.00, 500, saude);
         Produto livroJava = new ProdutoIsento("Use a Cabeça: Java", "Descomplicando Java",
-            112.90, 5, livros);
+            (decimal)112.90, 5, livros);
         Produto macbookPro = new ProdutoIsento("Macbook pro", "O melhor da marca!",
-            20000.00, 15, informatica);
+            (decimal)20000.00, 15, informatica);
 
         Console.WriteLine($"Produtos: \n\nCategoria: {vacina.Categoria.Nome} \nId: {vacina.Id} - {vacina.Nome}, {vacina.Descricao} R${vacina.Preco}" +
             $"\nQuantidade em estoque: {vacina.Estoque} \nImposto R${vacina.CalculaImposto()}");
