+++ Comex/Comex/Testes/TestaValidacaoDeProdutos.cs
@@ -8,7 +8,7 @@ namespace Comex.Testes
         Console.WriteLine("Criando Produto válido:");
         try
         {
-            Produto produto1 = new Produto("produto", "é um produto", 1.44, 20, moveis);
+            Produto produto1 = new Produto("produto", "é um produto", (decimal)1.44, 20, moveis);
             Console.WriteLine("Produto válido!");
         }
         catch (ArgumentException ex)
@@ -19,7 +19,7 @@ namespace Comex.Testes
         Console.WriteLine("\nCriando Produto com preço inválido:");
         try
         {
-            Produto produto1 = new Produto("produto", "é um produto", 0, 20, moveis);
+            Produto produto1 = new Produto("produto", "é um produto", (decimal)0, 20, moveis);
             Console.WriteLine("Produto válido!");
         }
         catch (ArgumentException ex)
@@ -30,7 +30,7 @@ namespace Comex.Testes
         Console.WriteLine("\nCriando Produto Isento com estoque inválido:");
         try
         {
-            Produto produto1 = new ProdutoIsento("produto", "é um produto", 1.44, -10, moveis);
+            Produto produto1 = new ProdutoIsento("produto", "é um produto", (decimal)1.44, -10, moveis);
             Console.WriteLine("Produto válido!");
         }
         catch (ArgumentException ex)
@@ -41,7 +41,7 @@ namespace Comex.Testes
         Console.WriteLine("\nCriando Produto Isento com categoria nula:");
         try
         {
-            Produto produto1 = new ProdutoIsento("produto", "é um produto", 1.44, 20, categoria: null);
+            Produto produto1 = new ProdutoIsento("produto", "é um produto", (decimal)1.44, 20, categoria: null);
             Console.WriteLine("Produto válido!");
         }
         catch (ArgumentException ex)
