+++ Comex/Comex/Testes/TestaVestuario.cs
@@ -8,7 +8,7 @@ namespace Comex.Testes
     {
         Categoria roupa = new Categoria("ROUPAS", "ATIVA");
 
-        Vestuario vestido = new Vestuario("Vestido rosa", "Bonito", 249.90, 100, roupa, "tamanho:M;cor:rosa");
+        Vestuario vestido = new Vestuario("Vestido rosa", "Bonito", (decimal)249.90, 100, roupa, "tamanho:M;cor:rosa");
 
         Console.WriteLine($"Produtos: \n\nCategoria: {vestido.Categoria.Nome} \nId: {vestido.Id} - {vestido.Nome}, {vestido.Descricao} " +
             $"R${vestido.Preco} \nQuantidade em estoque: {vestido.Estoque} \nImposto R${vestido.CalculaImposto()} \nCor: {vestido.Cor} - Tamanho: {vestido.Tamanho}");
